Revert "optimization release workflow (#205)" (#218)
This reverts commit aa3d52ccbb7538c388f5c822eb2ab0ee5956d52b.

+++ jenkins/pipelines/cd/optimization-build-tidb-linux-arm.groovy
@@ -12,7 +12,6 @@
 * @RELEASE_TAG
 * @PRE_RELEASE
 * @FORCE_REBUILD
-* @TIKV_PRID
 */
 def slackcolor = 'good'
 os = "linux"
@@ -20,7 +19,7 @@ arch = "arm64"
 platform = "centos7"
 
 def ifFileCacheExists(product,hash,binary) {
-    if (params.FORCE_REBUILD){
+    if (FORCE_REBUILD){
         return false
     }
     if(!fileExists("gethash.py")){
@@ -50,9 +49,6 @@ def TIDB_CTL_HASH = "master"
 def build_upload = { product, hash, binary ->
     stage("Build ${product}") {
         node("arm") {
-            if (ifFileCacheExists(product, hash, binary)) {
-                return
-            }
             def repo = "git@github.com:pingcap/${product}.git"
             def workspace = WORKSPACE
             dir("${workspace}/go/src/github.com/pingcap/${product}") {
@@ -164,32 +160,49 @@ try {
     stage("Build") {
         builds = [:]
 
-        builds["Build tidb-ctl"] = {
-            build_upload("tidb-ctl", TIDB_CTL_HASH, "tidb-ctl")
-        } 
-        builds["Build tidb"] = {
-            build_upload("tidb", TIDB_HASH, "tidb-server")
+        if (BUILD_TIKV_IMPORTER == "false") {
+            if (!ifFileCacheExists("tidb-ctl", TIDB_CTL_HASH, "tidb-ctl")) {
+                builds["Build tidb-ctl"] = {
+                build_upload("tidb-ctl", TIDB_CTL_HASH, "tidb-ctl")
+                } 
+            }
+            if (!ifFileCacheExists("tidb", TIDB_HASH, "tidb-server")) {
+                builds["Build tidb"] = {
+                build_upload("tidb", TIDB_HASH, "tidb-server")
+                }
+            }
+            if (!ifFileCacheExists("tidb-binlog", BINLOG_HASH, "tidb-binlog")) {
+                builds["Build tidb-binlog"] = {
+                build_upload("tidb-binlog", BINLOG_HASH, "tidb-binlog")
+                }
+            }
+            if (!ifFileCacheExists("tidb-tools", TOOLS_HASH, "tidb-tools")) {
+                builds["Build tidb-tools"] = {
+                build_upload("tidb-tools", TOOLS_HASH, "tidb-tools")
+                }
+            }
+            if (!ifFileCacheExists("pd", PD_HASH, "pd-server")) {
+                builds["Build pd"] = {
+                build_upload("pd", PD_HASH, "pd-server")
+                }
+            }
+            if (!ifFileCacheExists("ticdc", CDC_HASH, "ticdc")) {
+                builds["Build ticdc"] = {
+                build_upload("ticdc", CDC_HASH, "ticdc")
+                }
+            }
+            if (!ifFileCacheExists("br", BR_HASH, "br")) {
+                builds["Build br"] = {
+                build_upload("br", BR_HASH, "br")
+                }
+            }
+            if (!ifFileCacheExists("dumpling", DUMPLING_HASH, "dumpling")) {
+                builds["Build dumpling"] = {
+                build_upload("dumpling", DUMPLING_HASH, "dumpling")
+                }
+            }
         }
-        builds["Build tidb-binlog"] = {
-            build_upload("tidb-binlog", BINLOG_HASH, "tidb-binlog")
-        }
-        builds["Build tidb-tools"] = {
-            build_upload("tidb-tools", TOOLS_HASH, "tidb-tools")
-        }
-        builds["Build pd"] = {
-            build_upload("pd", PD_HASH, "pd-server")
-        }
-        builds["Build ticdc"] = {
-            build_upload("ticdc", CDC_HASH, "ticdc")
-        }
-        builds["Build br"] = {
-            build_upload("br", BR_HASH, "br")
-        }
-        builds["Build dumpling"] = {
-            build_upload("dumpling", DUMPLING_HASH, "dumpling")
-        }
-        
-        if (SKIP_TIFLASH == "false") {
+        if (SKIP_TIFLASH == "false" && BUILD_TIKV_IMPORTER == "false") {
             builds["Build tiflash"] = {
                 podTemplate(cloud: 'kubernetes-arm64', name: "build-arm-tiflash", label: "build-arm-tiflash",
                         instanceCap: 5, workspaceVolume: emptyDirWorkspaceVolume(memory: true),
@@ -271,34 +284,23 @@ try {
                     }
                     def target = "tikv-${RELEASE_TAG}-${os}-${arch}"
                     def filepath = "builds/pingcap/tikv/optimization/${TIKV_HASH}/centos7/tikv-server-${os}-${arch}.tar.gz"
-
-                    def specStr = "+refs/pull/*:refs/remotes/origin/pr/*"
-                    if (TIKV_PRID != null && TIKV_PRID != "") {
-                        specStr = "+refs/pull/${TIKV_PRID}/*:refs/remotes/origin/pr/${TIKV_PRID}/*"
-                    }
-                    def branch = TIKV_HASH
-                    if (RELEASE_BRANCH != null && RELEASE_BRANCH != "") {
-                        branch =RELEASE_BRANCH
-                    }
-
                     checkout changelog: false, poll: true,
-                            scm: [$class      : 'GitSCM', branches: [[name: "${branch}"]], doGenerateSubmoduleConfigurations: false,
+                            scm: [$class      : 'GitSCM', branches: [[name: "${TIKV_HASH}"]], doGenerateSubmoduleConfigurations: false,
                                   extensions  : [[$class: 'CheckoutOption', timeout: 30],
                                                  [$class: 'CloneOption', timeout: 60],
                                                  [$class: 'PruneStaleBranch'],
                                                  [$class: 'CleanBeforeCheckout']],
                                   submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'github-sre-bot-ssh',
-                                                                         refspec      : specStr,
+                                                                         refspec      : '+refs/heads/*:refs/remotes/origin/*',
                                                                          url          : 'git@github.com:tikv/tikv.git']]]
-                
-                    sh """
-                        git checkout -f ${TIKV_HASH}
-                        for a in \$(git tag --contains ${TIKV_HASH}); do echo \$a && git tag -d \$a;done
-                        git tag -f ${RELEASE_TAG} ${TIKV_HASH}
-                        git branch -D refs/tags/${RELEASE_TAG} || true
-                        git checkout -b refs/tags/${RELEASE_TAG}
-                    """
-                    
+                    if (BUILD_TIKV_IMPORTER == "false") {
+                        sh """
+                            for a in \$(git tag --contains ${TIKV_HASH}); do echo \$a && git tag -d \$a;done
+                            git tag -f ${RELEASE_TAG} ${TIKV_HASH}
+                            git branch -D refs/tags/${RELEASE_TAG} || true
+                            git checkout -b refs/tags/${RELEASE_TAG}
+                        """
+                    }
                     sh """
                         grpcio_ver=`grep -A 1 'name = "grpcio"' Cargo.lock | tail -n 1 | cut -d '"' -f 2`
                         if [[ ! "0.8.0" > "\$grpcio_ver" ]]; then
@@ -351,13 +353,14 @@ try {
                                                                                  url          : 'git@github.com:tikv/importer.git']]]
                         }
                     }
-                    sh """
-                        for a in \$(git tag --contains ${IMPORTER_HASH}); do echo \$a && git tag -d \$a;done
-                        git tag -f ${RELEASE_TAG} ${IMPORTER_HASH}
-                        git branch -D refs/tags/${RELEASE_TAG} || true
-                        git checkout -b refs/tags/${RELEASE_TAG}
-                    """
-                    
+                    if (BUILD_TIKV_IMPORTER == "false") {
+                        sh """
+                            for a in \$(git tag --contains ${IMPORTER_HASH}); do echo \$a && git tag -d \$a;done
+                            git tag -f ${RELEASE_TAG} ${IMPORTER_HASH}
+                            git branch -D refs/tags/${RELEASE_TAG} || true
+                            git checkout -b refs/tags/${RELEASE_TAG}
+                        """
+                    }
                     sh """
                         grpcio_ver=`grep -A 1 'name = "grpcio"' Cargo.lock | tail -n 1 | cut -d '"' -f 2`
                         if [[ ! "0.8.0" > "\$grpcio_ver" ]]; then
